requirements.txt
pandas
matplotlib
python-pptx
numpy
google-generativeai
scikit-learn
streamlit
